refactor: added import alias and did a whole lot of microfixes
